requirements.txt
ipykernel
jupyterlab
matplotlib
pandas
pillow
pytest
reportlab
